possible breaking change in Bindings's bind() and applying Bindings to each other:
it is now different from the semantics of the + and merge() operations.
The latter two produce a combination of the two sets of bindings.
Those have also been changed: now bindings in the left argument will
occur in the result, if the key appears in both, rather than the right.

The former are now defined to mean update the operand bindings by
making the operator binding's substitutions in all its right-hand-side
expressions. Bindings from the operator are not added to the operand.
There is also a special case for single variables, to allow
ww_load_session() to e.g. replace all occurrences of a bad "xhat"
variable by a good one, even on the bindings's left hand side.